Tidy up formatting

+++ src/main/java/com/sollace/coppergolem/client/CopperGolemEntityModel.java
@@ -1,5 +1,6 @@
 package com.sollace.coppergolem.client;
 
+import com.sollace.coppergolem.entity.CopperGolemEntity;
 import net.minecraft.client.model.ModelData;
 import net.minecraft.client.model.ModelPart;
 import net.minecraft.client.model.ModelPartBuilder;
@@ -10,8 +11,6 @@ import net.minecraft.client.render.entity.model.EntityModelPartNames;
 import net.minecraft.client.render.entity.model.SinglePartEntityModel;
 import net.minecraft.util.Hand;
 import net.minecraft.util.math.MathHelper;
-
-import com.sollace.coppergolem.entity.CopperGolemEntity;
 
 public class CopperGolemEntityModel extends SinglePartEntityModel<CopperGolemEntity> {
 
@@ -36,32 +35,32 @@ public class CopperGolemEntityModel extends SinglePartEntityModel<CopperGolemEnt
         this.leftArm = body.getChild(EntityModelPartNames.LEFT_ARM);
         this.rightLeg = root.getChild(EntityModelPartNames.RIGHT_LEG);
         this.leftLeg = root.getChild(EntityModelPartNames.LEFT_LEG);
-     }
+    }
 
     public static TexturedModelData getTexturedModelData() {
         ModelData data = new ModelData();
         ModelPartData root = data.getRoot();
 
         ModelPartData body = root.addChild(EntityModelPartNames.BODY, ModelPartBuilder.create()
-                .uv(0, 13).cuboid(-4, -7, -3, 8, 7, 5), ModelTransform.NONE);
+            .uv(0, 13).cuboid(-4, -7, -3, 8, 7, 5), ModelTransform.NONE);
 
         body.addChild(EntityModelPartNames.HEAD, ModelPartBuilder.create()
-                        .uv(36, 31).cuboid(-1, -1, -1.5F, 2, 2, 2)
-                        .uv(0, 0).cuboid(-4, -5, -4, 8, 5, 7)
-                        .uv(26, 16).cuboid(-1, -7, -1.5F, 2, 2, 2)
-                        .uv(0, 32).cuboid(-1.5F, -10, -2, 3, 3, 3), ModelTransform.pivot(0, -6, 0))
-                .addChild(EntityModelPartNames.NOSE, ModelPartBuilder.create()
-                        .uv(36, 26).cuboid(-1.5F, -1, 0, 2, 3, 2), ModelTransform.pivot(0.5F, -1, 3));
+                .uv(36, 31).cuboid(-1, -1, -1.5F, 2, 2, 2)
+                .uv(0, 0).cuboid(-4, -5, -4, 8, 5, 7)
+                .uv(26, 16).cuboid(-1, -7, -1.5F, 2, 2, 2)
+                .uv(0, 32).cuboid(-1.5F, -10, -2, 3, 3, 3), ModelTransform.pivot(0, -6, 0))
+            .addChild(EntityModelPartNames.NOSE, ModelPartBuilder.create()
+                .uv(36, 26).cuboid(-1.5F, -1, 0, 2, 3, 2), ModelTransform.pivot(0.5F, -1, 3));
         body.addChild(EntityModelPartNames.RIGHT_ARM, ModelPartBuilder.create()
-                .uv(26, 26).cuboid(0, -1, -2, 2, 10, 3), ModelTransform.pivot(4, -6, 0));
+            .uv(26, 26).cuboid(0, -1, -2, 2, 10, 3), ModelTransform.pivot(4, -6, 0));
         body.addChild(EntityModelPartNames.LEFT_ARM, ModelPartBuilder.create()
-                .uv(16, 26).cuboid(0, -1, -2, 2, 10, 3), ModelTransform.pivot(-6, -6, 0));
+            .uv(16, 26).cuboid(0, -1, -2, 2, 10, 3), ModelTransform.pivot(-6, -6, 0));
         root.addChild(EntityModelPartNames.RIGHT_LEG, ModelPartBuilder.create()
-                .uv(26, 9).cuboid(-2, 0, -3, 4, 3, 4)
-                .uv(23, 0).cuboid(-2, 3, -3, 4, 1, 5), ModelTransform.pivot(-2, 0, 0));
+            .uv(26, 9).cuboid(-2, 0, -3, 4, 3, 4)
+            .uv(23, 0).cuboid(-2, 3, -3, 4, 1, 5), ModelTransform.pivot(-2, 0, 0));
         root.addChild(EntityModelPartNames.LEFT_LEG, ModelPartBuilder.create()
-                .uv(0, 25).cuboid(-2, 0, -2, 4, 3, 4)
-                .uv(21, 20).cuboid(-2, 3, -2, 4, 1, 5), ModelTransform.pivot(2, 0, -1));
+            .uv(0, 25).cuboid(-2, 0, -2, 4, 3, 4)
+            .uv(21, 20).cuboid(-2, 3, -2, 4, 1, 5), ModelTransform.pivot(2, 0, -1));
 
         return TexturedModelData.of(data, 64, 64);
     }
@@ -79,12 +78,12 @@ public class CopperGolemEntityModel extends SinglePartEntityModel<CopperGolemEnt
             animationProgress = 1;
         }
 
-        root.yaw = -(float)Math.PI;
+        root.yaw = -(float) Math.PI;
         root.pivotY = 20;
 
-        float headSpinTime = (float)entity.getHeadSpinTime() / 10;
+        float headSpinTime = (float) entity.getHeadSpinTime() / 10;
 
-        float maxRotation = 2 * (float)Math.PI;
+        float maxRotation = 2 * (float) Math.PI;
 
         head.yaw = headYaw * 0.017453292F + MathHelper.lerp(headSpinTime, 0, maxRotation);
         head.pitch = headSpinTime > 0 ? 0 : -headPitch * 0.017453292F;
@@ -106,8 +105,8 @@ public class CopperGolemEntityModel extends SinglePartEntityModel<CopperGolemEnt
             leftLeg.pitch = 1.5F * MathHelper.wrap(limbAngle, 13.0F) * limbDistance;
 
             if (entity.isChasing()) {
-                rightArm.pitch = 1.25F + (float)MathHelper.sin(animationProgress / 2) / 10F;
-                leftArm.pitch = 1.25F + (float)MathHelper.cos(animationProgress / 2) / 10F;
+                rightArm.pitch = 1.25F + MathHelper.sin(animationProgress / 2) / 10F;
+                leftArm.pitch = 1.25F + MathHelper.cos(animationProgress / 2) / 10F;
             } else {
                 rightArm.pitch = 1.5F * MathHelper.wrap(limbAngle, 13) * limbDistance;
                 rightArm.yaw = -0.5F * MathHelper.wrap(limbAngle, 13) * limbDistance;
